infra/requirements.txt
-e .
pytest
aws_cdk.core
aws-cdk.aws-appsync==1.100.0
aws-cdk.aws-cloudwatch==1.100.0
aws-cdk.aws-codebuild==1.100.0
aws-cdk.aws-codecommit==1.100.0
aws-cdk.aws-codedeploy==1.100.0
aws-cdk.aws-codepipeline==1.100.0
aws-cdk.aws-codepipeline-actions==1.100.0 
aws-cdk.aws-dynamodb==1.100.0
aws-cdk.aws-iam==1.100.0
aws-cdk.aws-lambda==1.100.0
aws-cdk.aws-lambda-event-sources==1.32.2
aws-cdk.aws-logs==1.100.0
aws-cdk.aws-s3==1.100.0
aws-cdk.aws-s3-assets==1.100.0
aws-cdk.aws-s3-notifications==1.32.2
awscli==1.19.58
boto3==1.12.40
botocore==1.20.58
